requirements.txt
pyglet==1.3.2
opencv-python==3.4.0.12
numpy==1.14.3
nimblenet
PyYAML
neupy
sklearn
python-mnist
dill
keras
tensorflow
